+++ glsl.d.ts
@@ -1,0 +1,4 @@
+declare module "*.glsl" {
+  const value: string;
+  export default value;
+}
