+++ src/components/Animation/Animation.tsx
@@ -1,0 +1,19 @@
+import { Canvas } from "@react-three/fiber";
+
+import { Stats, OrbitControls } from "@react-three/drei";
+
+import Sphere from "./Sphere";
+
+function Animation() {
+  return (
+    <>
+      <Canvas>
+        <Sphere />
+        <OrbitControls />
+        <Stats />
+      </Canvas>
+    </>
+  );
+}
+
+export default Animation;
