+++ vite.config.js
@@ -1,8 +1,8 @@
 import { defineConfig } from "vite";
 
 import react from "@vitejs/plugin-react";
-import glsl from "vite-plugin-glsl";
+import viteTsconfigPaths from "vite-tsconfig-paths";
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [react(), glsl()],
+  plugins: [react(), viteTsconfigPaths()],
 });
